refactor / extract method
